Refactor helpers/error_hander/error_message.go
