+++ controllers/register.go
@@ -3,7 +3,7 @@ package controllers
 import (
 	"net/http"
 
-	errorHandler "github.com/gutakk/go-google-scraper/helpers/error_handler"
+	"github.com/gutakk/go-google-scraper/helpers/error_handler"
 	"github.com/gutakk/go-google-scraper/helpers/html"
 	session "github.com/gutakk/go-google-scraper/helpers/session"
 	"github.com/gutakk/go-google-scraper/models"
@@ -42,7 +42,7 @@ func (r *RegisterController) register(c *gin.Context) {
 	bindFormErr := c.ShouldBind(form)
 	if bindFormErr != nil {
 		for _, fieldErr := range bindFormErr.(validator.ValidationErrors) {
-			renderRegisterWithError(c, http.StatusBadRequest, errorHandler.ValidationErrorMessage(fieldErr), form)
+			renderRegisterWithError(c, http.StatusBadRequest, error_handler.ValidationErrorMessage(fieldErr), form)
 			return
 		}
 	}
