+++ helpers/error_handler/error_message.go
@@ -1,4 +1,4 @@
-package helpers
+package error_handler
 
 import (
 	"errors"
