+++ helpers/error_handler/error_message_test.go
@@ -1,4 +1,4 @@
-package helpers
+package error_handler
 
 import (
 	"errors"
